Legg på toggle på lagre barnehagebarn i KS
