+++ src/main/kotlin/no/nav/familie/ks/sak/barnehagelister/BarnehagelisteConsumer.kt
@@ -1,42 +1,52 @@
 package no.nav.familie.ks.sak.barnehagelister
 
-import no.nav.familie.kontrakter.felles.objectMapper
-import no.nav.familie.ks.sak.barnehagelister.domene.Barnehagebarn
+import no.nav.altinnkanal.avro.ReceivedMessage
+import no.nav.familie.ks.sak.barnehagelister.domene.BarnehagelisteMottatt
 import no.nav.familie.ks.sak.config.KafkaConfig
+import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.slf4j.LoggerFactory
 import org.springframework.context.annotation.Profile
 import org.springframework.kafka.annotation.KafkaListener
 import org.springframework.kafka.support.Acknowledgment
 import org.springframework.stereotype.Service
+import java.time.LocalDateTime
 
 @Service
-@Profile("!dev-postgres-preprod & !postgres")
+@Profile("!integrasjonstest & !dev-postgres-preprod & !postgres")
 class BarnehagelisteConsumer(
-    val barnehageBarnService: BarnehagebarnService,
+    val barnehageListeService: BarnehageListeService,
 ) {
     private val logger = LoggerFactory.getLogger(this::class.java)
 
     @KafkaListener(
-        id = "familie-ks-barnehagelister",
-        groupId = "familie-ks-barnehagelister-group",
-        topics = [KafkaConfig.BARNEHAGELISTE_TOPIC],
+        id = "familie-ks-sak-barnehageliste",
+        groupId = "familie-ks-sak-barnehageliste-group",
+        topics = [KafkaConfig.BARNEHAGELISTE_AAPEN_TOPIC],
+        containerFactory = "earliestConcurrentKafkaListenerContainerFactoryAvro",
     )
     fun listen(
-        message: String,
+        consumerRecord: ConsumerRecord<String, ReceivedMessage>,
         ack: Acknowledgment,
     ) {
-        val barnehagebarn: Barnehagebarn = objectMapper.readValue(message, Barnehagebarn::class.java)
-
-        logger.info("Barnehagebarn mottatt på kafka med id ${barnehagebarn.id}")
+        val data: ReceivedMessage = consumerRecord.value()
+        val key: String = data.archiveReference
+        logger.info("Barnehageliste mottatt på kafka med key $key")
 
         // Sjekk at vi ikke har mottat meldingen tidligere
-        if (barnehageBarnService.erBarnehageBarnMottattTidligere(barnehagebarn)) {
-            logger.info("Barnehagebarn med id ${barnehagebarn.id} er mottatt tidligere. Hopper over")
+        if (barnehageListeService.erBarnehagelisteMottattTidligere(key)) {
+            logger.info("Barnehageliste med key $key er mottatt tidligere. Hopper over")
             ack.acknowledge()
             return
         }
 
-        barnehageBarnService.lagreBarnehageBarn(barnehagebarn)
+        barnehageListeService.lagreBarnehagelisteMottattOgOpprettTaskForLesing(
+            BarnehagelisteMottatt(
+                meldingId = key,
+                melding = data.xmlMessage,
+                mottatTid = LocalDateTime.now(),
+            ),
+        )
+
         ack.acknowledge()
     }
 }
